digital/chat_bot/requirements.txt
line-bot-sdk
langchain==0.0.340
llama-index==0.9.7
google-api-python-client
google-cloud-storage
google-cloud-firestore